requirements.txt
absl-py==1.0.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
astroid==2.5.3
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.10.0
bleach==2.1.2
bm3d==3.0.9
Brlapi==0.6.6
captum==0.4.1
certifi==2018.1.18
cffi==1.14.6
chardet==3.0.4
click==8.0.3
cmake==3.18.4.post1
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cudarray==0.1.dev0
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.27
dataclasses==0.8
decorator==4.4.2
deeppy==0.1.dev0
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
Django==3.2.6
docker-pycreds==0.4.0
entrypoints==0.2.3.post1
ffmpeg-python==0.2.0
filelock==3.0.12
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
html5lib==0.999999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
image==1.5.33
image-slicer==2.1.1
imageio==2.9.0
importlib-metadata==3.10.0
importlib-resources==5.1.2
incremental==16.10.1
ipykernel==4.8.2
ipython==5.5.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isort==5.8.0
Jinja2==2.10
joblib==1.0.1
json5==0.9.5
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
jupyterlab==2.2.6
jupyterlab-server==1.2.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
libimagequant==2.14.1.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.3
MouseInfo==0.1.3
mpmath==1.2.1
munch==2.5.0
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
networkx==2.5.1
notebook==5.2.2
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
opencv-python==4.5.1.48
packaging==20.9
PAM==0.4.2
pandas==1.1.5
pandocfilters==1.4.2
pathtools==0.1.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.3.1
promise==2.3
prompt-toolkit==1.0.15
protobuf==3.19.4
psutil==5.9.0
pyarrow==6.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAutoGUI==0.9.53
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.2.0
PyGObject==3.26.1
pylint==2.7.4
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.0
PyScreeze==0.1.28
pyserial==3.4
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.8.2
python-debian==0.1.32
python3-xlib==0.15
pythran-openblas==0.3.6
pytorch-msssim==0.2.1
pytweening==1.0.4
pytz==2021.3
PyWavelets==1.1.1
pywhatkit==5.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
quantus==0.0.1
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
robust-loss-pytorch @ git+https://github.com/jonbarron/robust_loss_pytorch@9831f1db8006105fe7a383312fba0e8bd975e7f6
scikit-build==0.11.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sentry-sdk==1.5.5
service-identity==16.0.0
shortuuid==1.0.8
simplegeneric==0.8.1
simplejson==3.13.2
six==1.16.0
smmap==5.0.0
sos==4.1
soupsieve==2.3.1
sqlparse==0.4.1
ssh-import-id==5.7
system-service==0.3
systemd-python==234
termcolor==1.1.0
terminado==0.7
testpath==0.3.1
threadpoolctl==2.2.0
tifffile==2020.9.3
toml==0.10.2
torch==1.10.0
torch-dct==0.1.5
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==4.5.3
tqdm==4.61.2
traitlets==4.3.2
Twisted==17.9.0
typed-ast==1.4.3
typing_extensions==4.0.0
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.4.3
wadllib==1.3.2
Wand==0.6.7
wandb==0.12.10
wcwidth==0.1.7
webencodings==0.5
wget==3.2
wikipedia==1.4.0
wrapt==1.12.1
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
yaspin==2.1.0
zipp==3.4.1
zope.interface==4.3.2
